staging: android: ashmem: Changing return type from int to loff_t
Changing return type from int to loff_t. Actual return type of the
function (vfs_llseek) is loff_t (long long). Here due to implicit
converion from long long to int, result will be implementation defined.

+++ drivers/staging/android/ashmem.c
@@ -329,7 +329,7 @@ static ssize_t ashmem_read(struct file *file, char __user *buf,
 static loff_t ashmem_llseek(struct file *file, loff_t offset, int origin)
 {
 	struct ashmem_area *asma = file->private_data;
-	int ret;
+	loff_t ret;
 
 	if (asma->size == 0)
 		return -EINVAL;
